Channels working separately but when together, device is locked for p2p after sending lorawan data

+++ app/src/communication/lora/lora_device/lora_p2p_device/lora_p2p_device.h
@@ -52,7 +52,7 @@
 // Configures Lora Peer-to-Peer connection and starts listening if reception is enabled
 int setup_lora_p2p_connection(const struct device *lora_device);
 // Sends Lora Peer-to-Peer package and handles errors
-int send_lora_p2p_package(const struct device *lora_device, uint8_t *package, uint8_t package_size);
+int send_lora_p2p_package(const struct device *lora_device, uint8_t *package, int package_size);
 // // Acquires ownership of the LoRa device, changing the device configuration
 // int switch_ownership(LoraDevice lora_device);
